+++ R/qualpal-deprecated.R
@@ -1,0 +1,9 @@
+#' @title Deprecated functions in package \pkg{qualpal}.
+#'
+#' The functions listed below are deprecated and will be defunct in the next
+#' major version future. When possible, alternative functions with similar
+#' functionality are also mentioned. Help pages for deprecated functions are
+#' available at \code{help("<function>-deprecated")}.
+#' @name qualpal-deprecated
+#' @keywords internal
+NULL